requirements.txt
fastapi
uvicorn
gradio
librosa
pydub
sentence-transformers
spacy
PyPDF2
speechrecognition
pocketsphinx
scikit-learn
pyttsx3
vosk
pillow
reportlab
